+++ packages/at_lookup/lib/at_lookup.dart
@@ -5,5 +5,4 @@ export 'src/connection/outbound_connection_impl.dart';
 export 'src/transport/at_transport.dart';
 export 'src/exception/at_lookup_exception.dart';
 export 'src/cache/secondary_address_finder.dart';
-export 'src/cache/cacheable_secondary_address_finder.dart';
-export 'src/util/secure_socket_util.dart';
+export 'src/cache/proxy_secondary_address_finder.dart';