+++ packages/at_lookup/test/at_lookup_test_utils.dart
@@ -7,7 +7,7 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:at_chops/at_chops.dart';
-import 'package:at_lookup/at_lookup.dart';
+import 'package:at_lookup/at_lookup_io.dart';
 import 'package:at_lookup/src/connection/outbound_message_listener.dart';
 import 'package:mocktail/mocktail.dart';